requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
PyJWT
Flask-Mail
psycopg2-binary
python-dotenv
requests
gunicorn